dev: Adding interrupt generator for external interrupt emulation
This commit adds an interrupt generator. It owns an IntSourcePin.
There are two parameters: start and period. Start determines the
first time an interrupt is raised. An interrupt is raised by calling
raise() and lower() of the IntSourcePin in succession. If the period
is non zero, another interrupt is scheduled period ticks later.

Change-Id: I0a18b792a459b38bf168458377c44f58e63093b3